requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
alembic
redis
httpx
